Migrate to TypeScript.
