+++ .eslintrc.json
@@ -1,3 +1,3 @@
 {
-  "extends": "es/2015/server"
+  "extends": "es/node"
 }
