+++ package.json
@@ -14,17 +14,25 @@
     {
       "name": "Matthias Wagler",
       "email": "matthias.wagler@thenativeweb.io"
+    },
+    {
+      "name": "Hannes Leutloff",
+      "email": "hannes.leutloff@thenativeweb.io"
     }
   ],
-  "main": "lib/isolated.js",
+  "main": "build/lib/isolated.js",
+  "types": "build/lib/isolated.d.ts",
   "dependencies": {
-    "fs-extra": "7.0.1",
+    "fs-extra": "8.1.0",
     "recursive-readdir": "2.2.2",
     "temp": "0.9.0"
   },
   "devDependencies": {
-    "assertthat": "2.0.3",
-    "roboter": "4.0.4"
+    "@types/fs-extra": "8.0.0",
+    "@types/recursive-readdir": "2.2.0",
+    "@types/temp": "0.8.34",
+    "assertthat": "4.0.1",
+    "roboter": "7.1.1"
   },
   "repository": {
     "type": "git",
